refactor: move ecstore set heal helpers
