+++ crates/ecstore/src/layout/mod.rs
@@ -8,4 +8,5 @@ pub(crate) mod disks_layout;
 pub(crate) mod endpoint;
 pub(crate) mod endpoints;
 pub(crate) mod format;
+pub(crate) mod set_heal;
 pub(crate) mod set_layout;
